refactor(engine): Move credential resolution out of the engine credentials controller into a service (no-changelog)
