+++ packages/cli/src/modules/engine-v2/__tests__/engine-credentials.controller.test.ts
@@ -1,21 +1,13 @@
-import type { Logger } from '@n8n/backend-common';
 import type { Request, Response } from 'express';
-import type { IWorkflowExecuteAdditionalData } from 'n8n-workflow';
 import type { Mocked } from 'vitest';
 import { mock } from 'vitest-mock-extended';
 
-import type { CredentialsHelper } from '@/credentials-helper';
-import { CredentialNotFoundError } from '@/errors/credential-not-found.error';
 import { BadRequestError } from '@/errors/response-errors/bad-request.error';
 import { ForbiddenError } from '@/errors/response-errors/forbidden.error';
-import { NotFoundError } from '@/errors/response-errors/not-found.error';
-import type { CredentialsPermissionChecker } from '@/executions/pre-execution-checks';
 
 import type { ResolveCredentialRequest } from '../engine-credentials.contract';
 import { EngineCredentialsController } from '../engine-credentials.controller';
-
-const mocks = vi.hoisted(() => ({ getBase: vi.fn() }));
-vi.mock('@/workflow-execute-additional-data', () => ({ getBase: mocks.getBase }));
+import type { EngineCredentialsService } from '../engine-credentials.service';
 
 const resolveRequest: ResolveCredentialRequest = {
 	credential: { id: 'cred-1', name: 'Acme API', type: 'httpHeaderAuth' },
@@ -27,11 +19,8 @@ const resolveRequest: ResolveCredentialRequest = {
 const decrypted = { name: 'X-Api-Key', value: 'secret' };
 
 describe('EngineCredentialsController', () => {
-	let permissionChecker: CredentialsPermissionChecker;
-	let credentialsHelper: CredentialsHelper;
-	let logger: Logger;
+	let credentialsService: EngineCredentialsService;
 	let controller: EngineCredentialsController;
-	let additionalData: IWorkflowExecuteAdditionalData;
 
 	const newResponse = () => {
 		const res = { status: vi.fn(), json: vi.fn() };
@@ -42,23 +31,10 @@ describe('EngineCredentialsController', () => {
 	const newRequest = (body: unknown = resolveRequest) => ({ body }) as unknown as Request;
 
 	beforeEach(() => {
-		vi.resetAllMocks();
-		permissionChecker = mock<CredentialsPermissionChecker>();
-		credentialsHelper = mock<CredentialsHelper>();
-		logger = mock<Logger>();
-		controller = new EngineCredentialsController(
-			permissionChecker,
-			credentialsHelper,
-			mock<Logger>({ scoped: vi.fn().mockReturnValue(logger) }),
-		);
+		credentialsService = mock<EngineCredentialsService>();
+		controller = new EngineCredentialsController(credentialsService);
 
-		additionalData = {} as IWorkflowExecuteAdditionalData;
-		mocks.getBase.mockResolvedValue(additionalData);
-		vi.mocked(permissionChecker.findInaccessible).mockResolvedValue({
-			homeProject: mock(),
-			inaccessibleIds: [],
-		});
-		vi.mocked(credentialsHelper.getDecrypted).mockResolvedValue(decrypted);
+		vi.mocked(credentialsService.resolve).mockResolvedValue(decrypted);
 	});
 
 	describe('resolveCredential', () => {
@@ -71,38 +47,10 @@ describe('EngineCredentialsController', () => {
 			expect(res.json).toHaveBeenCalledExactlyOnceWith({ data: decrypted });
 		});
 
-		it('decrypts with the credential, mode and consumer node type from the request', async () => {
+		it('hands the validated body to the service', async () => {
 			await controller.resolveCredential(newRequest(), newResponse());
 
-			expect(credentialsHelper.getDecrypted).toHaveBeenCalledExactlyOnceWith(
-				additionalData,
-				{ id: 'cred-1', name: 'Acme API' },
-				'httpHeaderAuth',
-				'manual',
-				expect.objectContaining({
-					node: expect.objectContaining({ type: 'n8n-nodes-base.httpRequest' }),
-				}),
-			);
-		});
-
-		it('builds additional data for the execution in the request', async () => {
-			await controller.resolveCredential(newRequest(), newResponse());
-
-			expect(mocks.getBase).toHaveBeenCalledExactlyOnceWith({
-				userId: 'user-1',
-				workflowId: 'wf-1',
-				projectId: 'project-1',
-			});
-			// The engine's id, so `$execution.id` in a credential reads the right one.
-			expect(additionalData.executionId).toBe('exec-1');
-		});
-
-		it('checks that the workflow may use the credential before decrypting', async () => {
-			await controller.resolveCredential(newRequest(), newResponse());
-
-			expect(permissionChecker.findInaccessible).toHaveBeenCalledExactlyOnceWith('wf-1', [
-				'cred-1',
-			]);
+			expect(credentialsService.resolve).toHaveBeenCalledExactlyOnceWith(resolveRequest);
 		});
 
 		it.each([
@@ -123,41 +71,16 @@ describe('EngineCredentialsController', () => {
 			);
 			expect(res.status).not.toHaveBeenCalled();
 			// Unvalidated input must never reach the access check or the store.
-			expect(permissionChecker.findInaccessible).not.toHaveBeenCalled();
-			expect(credentialsHelper.getDecrypted).not.toHaveBeenCalled();
+			expect(credentialsService.resolve).not.toHaveBeenCalled();
 		});
 
-		it('refuses a credential the workflow may not use, and logs it', async () => {
-			vi.mocked(permissionChecker.findInaccessible).mockResolvedValue({
-				homeProject: mock(),
-				inaccessibleIds: ['cred-1'],
-			});
+		it('passes a service error on without writing a response', async () => {
+			const error = new ForbiddenError('Credential is not shared with the workflow');
+			vi.mocked(credentialsService.resolve).mockRejectedValue(error);
 			const res = newResponse();
 
-			await expect(controller.resolveCredential(newRequest(), res)).rejects.toThrow(ForbiddenError);
+			await expect(controller.resolveCredential(newRequest(), res)).rejects.toBe(error);
 			expect(res.status).not.toHaveBeenCalled();
-			expect(credentialsHelper.getDecrypted).not.toHaveBeenCalled();
-			// The client discards the body, so the operator's record is the log.
-			expect(logger.warn).toHaveBeenCalledExactlyOnceWith(
-				expect.stringContaining('Refused credential "cred-1" to workflow "wf-1"'),
-			);
-		});
-
-		it('answers 404 when the store has no credential with that id and type', async () => {
-			vi.mocked(credentialsHelper.getDecrypted).mockRejectedValue(
-				new CredentialNotFoundError('cred-1', 'httpHeaderAuth'),
-			);
-
-			await expect(controller.resolveCredential(newRequest(), newResponse())).rejects.toThrow(
-				NotFoundError,
-			);
-		});
-
-		it('passes on any other decryption error unchanged', async () => {
-			const error = new Error('cipher unavailable');
-			vi.mocked(credentialsHelper.getDecrypted).mockRejectedValue(error);
-
-			await expect(controller.resolveCredential(newRequest(), newResponse())).rejects.toBe(error);
 		});
 	});
 });
